requirements.txt
certifi==2023.7.22
grpcio==1.58.0
grpcio-tools==1.58.0
protobuf==4.24.3
